Added Custom Data
Added Custom Data to prepare a future update hability of the playlists.
